Add some color schemes and basic config
Our initial install of vim and the vimrc file was pretty basic, but that
was just enough to get rolling. Since that initial commit we've added
quite a bit to config so now is a good time to seal some of that away in
our vimrc and commit to version control (to share around).

- Use 'vimrc' rather than '.vimrc' so file is less invisible within .vim
  folder
- Add a bunch of initial config related to our first set of plugins
- Add some simple color schemes that are appealing (onedark by default)
- Ensure the readme file has notes about which plugins are installed:
  eventually we'll install via git submodules, but for now they need to
  be cloned manually