Set airline buffer list to show buffer numbers (and misc)
Vim airline shows a list of buffers at the top of the screen. Since
I've been doing a lot of buffer switching, it's nice to have the buffer
number printed there as well. In addition, make a couple of misc
changes to help with navigating vim for day to day use.

- Change some settings related to the buffer list displayed by airline
- Also install the "silver-searcher" plugin to make use of AG in vim
- Uncomment the section related to disabling arrow keys in insert mode;
  I'm still not 100% ready to esc out to normal mode to move back a few
- Map some keys for changing splits/buffers a little quicker
- Disable netrw (using nerdtree instead) and turn off default wrapping
- Disable the mustache abbreviations: they're not working right yet...