Change vim plugin config and clean zshrc a bit
In an effort to slim down some of my dependencies (and optimize the vim
dotfiles in particular) I was goind some experimentation this weekend. I
switched off of Ctrl-P (which had some ruby deps) on to FZF instead
(which is more recent and maintained) and went from powerline to
lightline. Aside from that, I was able to clear out some unwanted paths
from my zshrc.

- Add tmux-powerline config to the repo (as is currently)
- Update vim-plug kept within the repo
- Unmap another key combo that fugitive seems to set up
- Switch over to FZF rather than ctrl-p, and set up config in vimrc,
  specifically ensuring that the fuzzy finder WILL see hidden files, but
  will also ignore anything gitignored.
- Change some of the configuration stored in the plugin file around to
  better accomodate javascript syntax (etc)
- Turn on linenumbers for "hybrid" line number mode in vim
- Tweak some of the settings related to tab stops / autoindent
- Also apply some minor tweaks to zshrc to keep the path as lean as
  possible, but adding the homebrew path as necessary