Update kitty configuration and fix vim settings
I'm using Kitty terminal now, but I've noticed on OSX that there are
font rendering issues if I don't update the default terminfo. Add to the
instructions and include scripts to make it work.

Somehow vim also had a weird set of characters get commited to the
keybindings, which was causing weird errors.  YouCompleteMe is also not
really working for me as much.